requirements.txt
pyTelegramBotAPI==4.12.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
gspread==5.10.0 